added custom Exceptions

+++ src/main/java/com/javaAssignment/Task2/api/controller/BookController.java
@@ -1,5 +1,7 @@
 package com.javaAssignment.Task2.api.controller;
 
+import com.javaAssignment.Task2.api.customException.BookAlreadyExistsException;
+import com.javaAssignment.Task2.api.customException.NoBookFoundException;
 import com.javaAssignment.Task2.entity.Books;
 import com.javaAssignment.Task2.service.BookService;
 import org.springframework.http.HttpStatus;
@@ -8,6 +10,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @RestController
 public class BookController {
@@ -30,22 +33,32 @@ private final BookService bookService;
     }
     @PostMapping("/admin/books")
     public ResponseEntity<String> addBook(@RequestBody Books book) {
-        if (bookService.doesBookExist(book.getTitle())) {
-            return ResponseEntity.status(HttpStatus.CONFLICT)
-                    .body("Book with title '" + book.getTitle() + "' already exists in the database.");
+        try {
+            if (bookService.doesBookExist(book.getTitle())) {
+                throw new BookAlreadyExistsException("Book with title '" + book.getTitle() + "' already exists in the database.");
+            }
+            bookService.addBook(book);
+            return ResponseEntity.status(HttpStatus.CREATED)
+                    .body("Book added successfully.");
+        } catch (BookAlreadyExistsException ex) {
+            return ResponseEntity.status(HttpStatus.CONFLICT).body(ex.getMessage());
         }
-        bookService.addBook(book);
-        return ResponseEntity.status(HttpStatus.CREATED)
-                .body("Book added successfully.");
     }
     @DeleteMapping("/admin/books/{id}")
     public ResponseEntity<String> deleteBook(@PathVariable Long id) {
-        boolean isDeleted = bookService.deleteBookById(id);
-        if (isDeleted) {
+        try{
+            Optional<Books> book = bookService.getBookById(id);
+            if(book.isEmpty()){
+                throw new NoBookFoundException("The Book with the id: "+ id + " not found.");
+            }
+            boolean isDeleted = bookService.deleteBookById(id);
             return ResponseEntity.ok("Book with ID " + id + " has been deleted.");
-        } else {
+        }
+        catch (NoBookFoundException ex)
+        {
             return ResponseEntity.status(404).body("Book with ID " + id + " not found.");
         }
+
     }
     @GetMapping("/admin/books/search")
     public ResponseEntity<?> searchBooksByTitle(@RequestParam("title") String title) {
